+++ packages/core/types/src/http/common/request.ts
@@ -1,8 +1,8 @@
 export interface SelectParams {
   /**
-   * The fields and relations to retrieve.
+   * The fields and relations to retrieve separated by commas.
    * 
-   * Learn more in the [API reference](https://docs.medusajs.com/v2/api/store#select-fields-and-relations).
+   * Learn more in the [API reference](https://docs.medusajs.com/api/store#select-fields-and-relations).
    */
   fields?: string
 }