+++ packages/modules/cache-inmemory/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/cache-inmemory
 
+## 2.0.1
+
 ## 2.0.0
 
 ### Major Changes