+++ packages/modules/event-bus-local/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/event-bus-local
 
+## 2.0.1
+
 ## 2.0.0
 
 ### Major Changes