+++ packages/modules/providers/fulfillment-manual/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/fulfillment-manual
 
+## 2.0.1
+
 ## 2.0.0
 
 ### Major Changes