+++ packages/modules/region/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/region
 
+## 2.0.1
+
 ## 2.0.0
 
 ### Major Changes